jeff_agent/pyproject.toml
[project]
name = "jeff-agent"
version = "0.1.0"
description = "LangChain/LangGraph participant agent for A2A badminton scheduling"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.11",
    "langchain>=1.0.5",
    "langchain-google-genai>=3.0.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.30.0",
]
